fix: ensure no failure during detached HEAD state
when doing a rebase, we enter a detached HEAD state, which causes
`git symbolic-ref` to fail. as a fallback, we try to pull the
information from the commit message. presumably, if we're doing a git
rebase, this hook as already gone and put the footer in place. we try
to grab that however we may