fix: suppress stderr when we can't get tkt from branch
in the event of an interactive rebase where we enter the detached HEAD
state:
```
r SHA1 msg1
f SHA2 msg2
```
getting the ticket name from the branch fails. while we handle this
gracefully, stderr is piped out, which can look confusing with a
successful rebase:
```
git-test git:(dx-1000/hello) git rebase -i HEAD~2
fatal: ref HEAD is not a symbolic ref
[detached HEAD 08d6e4c] world 2
 Date: Sun Feb 2 16:28:37 2025 -0500
 1 file changed, 2 insertions(+)
```

pipe the output up to the parent process to suppress. note we do not
ignore here, as we don't want to suppress stdout. we could arguably use
array notation for each stream, but we'll get there when we need to